Implement complete authentication and database-backed shopping cart
- Added ASP.NET Core Identity with Entity Framework Core
- Created AppUser model extending IdentityUser with custom properties
- Implemented database-backed cart with CartItem entity
- Added user-specific cart persistence and isolation
- Fixed LINQ queries for EF Core compatibility (case-insensitive matching)
- Integrated authentication UI with login/register/logout functionality
- Updated navigation to show auth-aware links and cart access
- Refactored ShoppingCartController to use async/await patterns
- Added proper database migrations and context configuration
- Cart items now persist across application restarts
- Each authenticated user has their own isolated cart
- Updated views to work with new CartItem model
- All cart operations (add/remove/view) working with database storage

+++ ch1/src/TightlyCoupled.WebShop/Services/PaymentProcessor.cs
@@ -5,7 +5,7 @@ namespace TightlyCoupled.WebShop.Services
     /// </summary>
     public class PaymentProcessor
     {
-        public string ProcessPayment(double amount)
+        public string ProcessPayment(decimal amount)
         {
             // Simulated payment processing logic
             if (amount > 1000)
